Connect posthog to sentry
